start moving common functions from doplay to a utils module
