+++ crack-core/src/commands/music/doplay.rs
@@ -3,12 +3,12 @@ use super::doplay_utils::insert_track;
 use super::doplay_utils::queue_keyword_list;
 
 use crate::commands::doplay_utils::queue_keyword_list_w_offset;
+use crate::commands::get_call_with_fail_msg;
 use crate::{
     commands::skip::force_skip_top_track,
-    connection::get_voice_channel_for_user,
     errors::{verify, CrackedError},
     guild::settings::{GuildSettings, DEFAULT_PREMIUM},
-    handlers::{track_end::update_queue_messages, IdleHandler, TrackEndHandler},
+    handlers::track_end::update_queue_messages,
     http_utils,
     interface::create_now_playing_embed,
     messaging::{
@@ -40,15 +40,12 @@ use ::serenity::{
         EditMessage,
     },
 };
-use poise::{
-    serenity_prelude::{self as serenity, Attachment, Http},
-    CreateReply,
-};
+use poise::serenity_prelude::{self as serenity, Attachment, Http};
 use reqwest::Client;
 use songbird::{
     input::{AuxMetadata, Compose, HttpRequest, Input as SongbirdInput, YoutubeDl},
     tracks::TrackHandle,
-    Call, Event, TrackEvent,
+    Call,
 };
 use std::process::Stdio;
 use std::{
@@ -57,7 +54,7 @@ use std::{
     error::Error as StdError,
     path::Path,
     process::Output,
-    sync::{atomic::AtomicBool, Arc},
+    sync::Arc,
     time::Duration,
 };
 use tokio::process::Command;
@@ -182,106 +179,6 @@ fn get_msg(mode: Option<String>, query_or_url: Option<String>, is_prefix: bool) 
         step1
     }
 }
-
-/// Get the call handle for songbird.
-// FIXME: Does this need to take the GuildId?
-pub async fn get_call_with_fail_msg(
-    ctx: Context<'_>,
-    guild_id: serenity::GuildId,
-) -> Result<Arc<Mutex<Call>>, Error> {
-    let manager = songbird::get(ctx.serenity_context()).await.unwrap();
-    match manager.get(guild_id) {
-        Some(call) => Ok(call),
-        None => {
-            // try to join a voice channel if not in one just yet
-            //match summon_short(ctx).await {
-            // TODO: Don't just return an error on failure, do something smarter.
-            let channel_id = {
-                let guild = ctx.guild().ok_or(CrackedError::NoGuildCached)?;
-                get_voice_channel_for_user(&guild.clone(), &ctx.author().id)?
-            };
-            match manager.join(guild_id, channel_id).await {
-                Ok(call) => {
-                    let text = set_global_handlers(ctx, call.clone(), guild_id, channel_id).await?;
-
-                    let msg = ctx
-                        .send(CreateReply::default().content(text).ephemeral(true))
-                        .await?
-                        .into_message()
-                        .await?;
-                    ctx.data().add_msg_to_cache(guild_id, msg);
-                    // }
-                    Ok(call)
-                    // Ok(manager.get(guild_id).unwrap())
-                },
-                Err(_) => {
-                    // FIXME: Do something smarter here also.
-                    let embed = CreateEmbed::default()
-                        .description(format!("{}", CrackedError::NotConnected));
-                    send_embed_response_poise(ctx, embed).await?;
-                    Err(CrackedError::NotConnected.into())
-                },
-            }
-        },
-    }
-}
-
-/// Set the global handlers.
-pub async fn set_global_handlers(
-    ctx: Context<'_>,
-    call: Arc<Mutex<Call>>,
-    guild_id: GuildId,
-    channel_id: ChannelId,
-) -> Result<String, Error> {
-    let data = ctx.data();
-
-    let manager = songbird::get(ctx.serenity_context())
-        .await
-        .ok_or(CrackedError::NoSongbird)?;
-
-    let mut handler = call.lock().await;
-    // unregister existing events and register idle notifier
-    handler.remove_all_global_events();
-
-    let guild_settings_map = data.guild_settings_map.read().unwrap().clone();
-
-    let _ = guild_settings_map.get(&guild_id).map(|guild_settings| {
-        let timeout = guild_settings.timeout;
-        if timeout > 0 {
-            let premium = guild_settings.premium;
-            handler.add_global_event(
-                Event::Periodic(Duration::from_secs(5), None),
-                IdleHandler {
-                    http: ctx.serenity_context().http.clone(),
-                    manager: manager.clone(),
-                    channel_id,
-                    guild_id: Some(guild_id),
-                    limit: timeout as usize,
-                    count: Default::default(),
-                    no_timeout: Arc::new(AtomicBool::new(premium)),
-                },
-            );
-        }
-    });
-
-    handler.add_global_event(
-        Event::Track(TrackEvent::End),
-        TrackEndHandler {
-            guild_id,
-            http: ctx.serenity_context().http.clone(),
-            call: call.clone(),
-            data: ctx.data().clone(),
-        },
-    );
-
-    let text = CrackedMessage::Summon {
-        mention: channel_id.mention(),
-    }
-    .to_string();
-
-    Ok(text)
-}
-
 /// Sends the searching message after a play command is sent.
 /// Also defers the interaction so we won't timeout.
 async fn send_search_message(ctx: Context<'_>) -> Result<Message, Error> {
